requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
scikit-learn==1.8.0
pandas==2.0.3
numpy==1.24.3
groq==0.30.0
python-dotenv==1.0.0
